+++ src/styles/switchSpanStyle.ts
@@ -10,9 +10,8 @@ const switchSpanStyle = css`
   left: 0;
   right: 0;
   bottom: 0;
-  border: 2px solid var(--color-text);
   border-radius: 34px;
-  /* background-color: var(--color-background-weak); */
+  background-color: var(--color-background-switch);
   transition: 0.4s;
 
   &::before {
@@ -22,7 +21,6 @@ const switchSpanStyle = css`
     width: 26px;
     left: 4px;
     bottom: 4px;
-    border: 2px solid var(--color-text);
     border-radius: 50%;
     background-color: var(--color-background);
     transition: 0.4s;